chore(ts/e2e): refactored json-schema-to-zod e2e tests into unified ts/e2e-tests location
