+++ ts/e2e-tests/runtimes/node/json-schema-to-zod-v4/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@e2e-tests/node-json-schema-to-zod-v4",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "typecheck": "tsc --noEmit",
+    "test:e2e": "bun test e2e.test.ts",
+    "test:e2e:node": "bun test e2e.test.ts"
+  },
+  "dependencies": {
+    "@composio/json-schema-to-zod": "workspace:*",
+    "zod": "^4.1.0"
+  },
+  "devDependencies": {
+    "@e2e-tests/utils": "workspace:*"
+  }
+}
