--- ts/packages/json-schema-to-zod/test/e2e/.gitignore
@@ -1,1 +1,0 @@
-node_modules/
