--- ts/packages/json-schema-to-zod/test/e2e/zod-v3/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@test-e2e/json-schema-to-zod-e2e-zod-v3",
-  "version": "1.0.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "test": "vitest run"
-  },
-  "dependencies": {
-    "@composio/json-schema-to-zod": "workspace:*",
-    "zod": "3.25.76",
-    "zod-to-json-schema": "catalog:"
-  },
-  "devDependencies": {
-    "vitest": "catalog:"
-  }
-}
